Moved 'Query' folder from Witsml project to WitsmlExplorer.Api
