+++ Src/WitsmlExplorer.Api/Query/LogQueries.cs
@@ -4,7 +4,7 @@ using Witsml.Data;
 using Witsml.Data.Curves;
 using Witsml.Extensions;
 
-namespace Witsml.Query
+namespace WitsmlExplorer.Api.Query
 {
     public static class LogQueries
     {
