+++ Src/WitsmlExplorer.Api/Query/WellboreQueries.cs
@@ -1,7 +1,7 @@
 using Witsml.Data;
 using Witsml.Extensions;
 
-namespace Witsml.Query
+namespace WitsmlExplorer.Api.Query
 {
     public static class WellboreQueries
     {
@@ -74,7 +74,5 @@ namespace Witsml.Query
                 }.AsSingletonList()
             };
         }
-
-
     }
 }
