+++ Src/WitsmlExplorer.Api/Workers/BatchModifyWellWorker.cs
@@ -8,6 +8,7 @@ using Witsml.Data;
 using Witsml.Extensions;
 using WitsmlExplorer.Api.Jobs;
 using WitsmlExplorer.Api.Models;
+using WitsmlExplorer.Api.Query;
 using WitsmlExplorer.Api.Services;
 
 namespace WitsmlExplorer.Api.Workers
@@ -29,8 +30,8 @@ namespace WitsmlExplorer.Api.Workers
         public async Task<(WorkerResult, RefreshAction)> Execute(BatchModifyWellJob job)
         {
             Verify(job.Wells);
-            var queries = job.Wells.Select(well => CreateUpdateQuery(well));
-            var UpdateWellTasks = queries.Select(q => witsmlClient.UpdateInStoreAsync(q));           
+            var wellsToUpdate = job.Wells.Select(well => WellQueries.UpdateWitsmlWell(well));
+            var UpdateWellTasks = wellsToUpdate.Select(wellToUpdate => witsmlClient.UpdateInStoreAsync(wellToUpdate));           
 
             Task resultTask = Task.WhenAll(UpdateWellTasks);
             await resultTask;
@@ -51,22 +52,6 @@ namespace WitsmlExplorer.Api.Workers
         private void Verify(IEnumerable<Well> wells)
         {
             if (!wells.Any()) throw new InvalidOperationException("payload cannot be empty");
-        }
-
-        private static WitsmlWells CreateUpdateQuery(Well well)
-        {
-            return new WitsmlWells
-            {
-                Wells = new WitsmlWell
-                {
-                    Uid = well.Uid,
-                    Name = well.Name,
-                    Field = well.Field,
-                    TimeZone = well.TimeZone,
-                    Country = well.Country,
-                    Operator = well.Operator
-                }.AsSingletonList()
-            };
         }
     }
 }
