+++ Src/WitsmlExplorer.Api/Workers/ModifyWellboreWorker.cs
@@ -4,10 +4,10 @@ using System.Threading.Tasks;
 using Serilog;
 using Witsml;
 using Witsml.Extensions;
-using Witsml.Query;
 using Witsml.ServiceReference;
 using WitsmlExplorer.Api.Jobs;
 using WitsmlExplorer.Api.Models;
+using WitsmlExplorer.Api.Query;
 using WitsmlExplorer.Api.Services;
 
 namespace WitsmlExplorer.Api.Workers
